Fix undo for spell check.
It was broken because it was calling reformatAll after the changes,
outside of any undo recorded edit steps.

The fix is to not delay reformatting after all replaces have been done,
but reformat each changed paragraph after each change, which nicely
works and records any reformatting in the SinglePara undo object we're
already creating.

Since reformatting can cause words to move around, also stop using
searchLine, and start using the normal mark to highlight which word we're
about to replace.