Fix tab visual glitch on high dpi as good as possible
The tab should melt with the view below it, by cutting a hole into the separator that separates the tab bar from the widget below.

This was achieved by drawing the tab background over the separator line, covering it.

Also, the tab shouldn't have a bottom border itself, obviously.

This was achieved by drawing a full polygon with full borders for the tab background, but limiting it by a clipping region, so the bottom border was cut off (as it was drawn outside the boundaries of the clipping region). This broke on high dpi screens, as the sizes for the clipping box and the tab were calculated on pixel basis, and didn't line up with higher precision. Therefore, you could see the bottom line of the tab background.

Making the clipping region smaller, however, made the background not cover the seperator line anymore.

This commit fixes this, by drawing the background first as a polygon (without borders) and then draw the borders (without the bottom border). The borders will be drawn with a smaller clipping region, so they don't cross the seperator line.