Test text import
Test doesn't work, but that's not because the test is broken – the import is 😂

To be fair, the test also is broken: It compares the imported stuff case-sensitive, while for some things, text file import can't preserve case.

But some lines are imported wrong by the import. I'm gonna investigate this!