Remove tox.ini
I thought about keeping it, as it still allows running tests in multiple environments locally. But my fear is, that if it is not run in CI, it will only be used rarely, never tested, and very likely to get outdated and cause trouble when someone tries to use it.