Pass reportlabs canvas to PDFDrawOp
Passing the canvas additionally to the output string, we can smoothly transition all operations from rendering to the output string to rendering to the reportlab canvas.

A smooth transitions allows to test the effects of each change individually.

Of course, we will later-on remove  the output string parameter when it's not needed anymore.