Fix installation (e. g. to "/opt" or a custom location)
Depending on how trelby is installed, the `src` directory will be copied to different places. The `bin/trelby` script needs to know about that location, so it can start the application. That's why the installation process in `setup.py` rewrites the path that 'bin/trelby' uses to import the `src` files.

So no matter whether you create e. g. an rpm using `setup.py bdist_rpm` or install Trelby to `/opt` using `setup.py install`, it should have the right location.

But this behaviour was broken for the `install` command:

When using the `install` command, setup.py will first create an egg in a build dir and make all the installation steps in there, as if this was the final installation. Only then, the `easy_install` step copies the .egg to the final location (e. g. `/opt`).

Because of this, the code that rewrote the import path worked with a relative path inside the build directory, which of course was invalid in the final installation directory.

To get the final final installation directory, I needed to hook into the `easy_install` command and do the rewrite there.

Probably this broke because distutils changed the way the `install` step works by introducing `easy_install`.

After this commit, it works reliably for `setup.py install` to any location and `setup.py bdist_rpm` continues to work as expected. I couldn't test builiding debs as I don't get how to install the dependencies for this, but as rpms work I assume this is also going to work.